requirements.txt
mailjet_rest>=1.3.3
python-dotenv
selenium>=4.7.2
beautifulsoup4>=4.11.1
requests>=2.28.1
pydub>=0.25.1
SpeechRecognition>=3.9.0
